Repurpose the profile-validation test file instead of deleting it
The prior fix pass deleted tests/server/subscription/test_profile_validation.py,
a pre-existing file whose tests only exercised the now-removed
ConfigResolver._resolve_profile. The quality gate builds its changed-file list
from git diff against origin/main, which lists a deleted path, then runs ruff,
ruff format, and coverage on it. All three errored on the missing file while the
core suite was green.

Keep the file at its path and give it the live resolver-level profile-name
validation tests (unknown name and non-string both raise ConfigurationError),
moving them out of test_config_resolver.py so there is no duplication.
